--- a/ci/configs/ci.env
+++ b/ci/configs/ci.env
@@ -1,0 +1,38 @@
+# CI Environment Configuration
+# These variables are used across all CI containers
+
+# Node.js Configuration
+NODE_ENV=test
+NODE_OPTIONS=--max-old-space-size=4096
+
+# CI Flags
+CI=true
+FORCE_COLOR=1
+TERM=xterm-256color
+
+# pnpm Configuration
+PNPM_HOME=/pnpm
+PNPM_STORE_DIR=/pnpm-store
+
+# Telemetry Opt-out
+ASTRO_TELEMETRY_DISABLED=1
+NEXT_TELEMETRY_DISABLED=1
+GATSBY_TELEMETRY_DISABLED=1
+
+# Test Configuration
+TEST_TIMEOUT=30000
+VITEST_REPORTER=verbose
+
+# Playwright Configuration
+PLAYWRIGHT_BROWSERS_PATH=/ms-playwright
+PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD=0
+
+# Build Configuration
+BUILD_CACHE_DIR=/build-cache
+TURBO_CACHE_DIR=/turbo-cache
+
+# Security
+NODE_TLS_REJECT_UNAUTHORIZED=1
+
+# Performance
+UV_THREADPOOL_SIZE=16