… .github/workflows/README.md
… .github/workflows/docker-cleanup.yml
… .github/workflows/docker-images.yml
… .github/workflows/docker-pr-builds.yml
… .github/workflows/docker-security-updates.yml
… .github/workflows/e2e-sharded.yml
… .github/workflows/nightly-tests.yml
… ci/IMPLEMENTATION_COMPLETE.md
… ci/PHASE2_BUILD_DEPLOY_IMAGE_SUMMARY.md
… ci/PHASE2_SUMMARY.md
… ci/PHASE2_TEST_IMAGE_SUMMARY.md
… ci/PHASE4_IMAGE_CICD_SUMMARY.md
… ci/README.md
… ci/build/docker-compose.yml
… ci/build/scripts/help.sh
… ci/configs/github-e2e-workflow.yml
… ci/configs/github-workflow-example.yml
… ci/docker-compose.yml
… ci/docs/BUILD_PROCESS.md
… ci/docs/CI_CD_PIPELINE.md
… ci/docs/DEPLOY_IMAGE.md
… ci/docs/E2E_SHARDED_CONTAINERIZATION.md
… ci/docs/E2E_SHARDED_QUICK_REFERENCE.md
… ci/docs/NIGHTLY_TESTS_QUICK_REFERENCE.md
… ci/docs/PHASE3_E2E_SHARDED_SUMMARY.md
… ci/docs/PHASE3_NIGHTLY_TESTS_SUMMARY.md
… ci/docs/PHASE3_PR_TESTS_MIGRATION.md
… ci/docs/PR_TESTS_QUICK_REFERENCE.md
… ci/docs/QUICK_REFERENCE.md
… ci/docs/TEST_IMAGE.md