## Summary
- Implemented comprehensive CI/CD containerization strategy achieving 40-75% performance improvements
- Migrated all GitHub Actions workflows to Docker-based architecture
- Established GitHub Container Registry (ghcr.io) for image management

## Performance Improvements Achieved
| Workflow | Before | After | Improvement |
|----------|--------|-------|-------------|
| PR Tests | 3m 13s | 1m 56s | **40% faster** |
| E2E Sharded | 2m 42s | 1m 26s | **47% faster** |
| Nightly Tests | 2m 0s | 30s | **75% faster** |

## Key Changes

### Phase 1: Docker Registry & Baselines
- Set up GitHub Container Registry integration
- Created performance measurement scripts
- Established baseline metrics for comparison

### Phase 2: Docker Images & CI/CD Pipeline
- Created multi-stage Docker images:
  - Base image with Node.js and system dependencies
  - Test image with Playwright browsers
  - Deploy image optimized for production
- Implemented automated image building and caching

### Phase 3: Workflow Migration
- Migrated all CI/CD workflows to use Docker containers
- Optimized caching strategies for maximum performance
- Implemented parallel execution where possible

### Documentation
- Comprehensive guides in `ci/docs/`:
  - Quick reference guides for each workflow
  - Detailed implementation documentation
  - Performance analysis reports
- Updated CLAUDE.md with containerization information

## Test Plan
- [x] All existing tests pass in containerized environment
- [x] Performance targets met or exceeded
- [x] Docker images build successfully
- [x] Workflows execute correctly with new architecture
- [x] Documentation is complete and accurate

Closes #141